Restructuring to try to access animation functions as methods (incomplete)
